backend/requirements.txt
fastapi[standard]==0.115.0

uvicorn==0.32.0
python-multipart==0.0.12
pydantic==2.9.2
opencv-python==4.10.0.84
numpy==2.2.0
torch==2.10.0
torchvision==0.25.0
pillow==11.0.0
scikit-image==0.24.0
albumentations==1.4.20
matplotlib==3.9.2
pandas==2.2.3
reportlab==4.2.5
loguru==0.7.2
ultralytics